Refactored the days enum into a settings directory
